requirements.txt
# Core
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.6.0
flask-mail==0.9.1
flask-cors==4.0.0
resend==2.0.0

# Database
psycopg[binary]>=3.1.0
supabase==2.3.0

# Utils
python-dotenv==1.0.0
cloudinary==1.36.0
blinker==1.7.0

# Document generation
python-docx==0.8.11
google-api-python-client==2.70.0
google-auth==2.16.0
google-auth-oauthlib==0.8.0
google-auth-httplib2==0.1.0

# Document generation
python-docx==0.8.11
google-api-python-client==2.70.0
google-auth-oauthlib==0.8.0
google-auth-httplib2==0.1.0
cachetools==5.3.0

# Whiteboard
flask-socketio==5.3.4
eventlet==0.33.3

# Server
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
